Added ignore attributes to properties that do not have to be serialized.
